03.09.26

+++ Assets/Project/Scripts/DialogueSystem/Phrase/Phrase.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public struct Phrase
 {
     [field: SerializeField] public PhraseInfo Info;
-    [field: SerializeField] public int Delay;
+    [field: SerializeField] public float Delay;
 }
